git-svn-id: svn://svn.handbrake.fr/HandBrake/branches/opencl@5264 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ win/CS/HandBrake.Interop/HandBrakeInterop/SourceData/Title.cs
@@ -140,7 +140,7 @@ namespace HandBrake.Interop.SourceData
 		{
 			get
 			{
-				return (int)Math.Ceiling(((double)this.Duration.TotalSeconds) * this.Framerate);
+				return (int)Math.Ceiling(this.Duration.TotalSeconds * this.Framerate);
 			}
 		}
